2nd Week: Create Converter component
